+++ plugins/Goals/API.php
@@ -23,6 +23,7 @@ use Piwik\Plugins\API\DataTable\MergeDataTables;
 use Piwik\Plugins\CoreHome\Columns\Metrics\ConversionRate;
 use Piwik\Plugins\Goals\Columns\Metrics\AverageOrderRevenue;
 use Piwik\Plugin\ReportsProvider;
+use Piwik\Segment;
 use Piwik\Segment\SegmentExpression;
 use Piwik\Site;
 use Piwik\Tracker\Cache;
@@ -54,6 +55,7 @@ use Piwik\Validators\WhitelistedValue;
 class API extends \Piwik\Plugin\API
 {
     const AVG_PRICE_VIEWED = 'avg_price_viewed';
+    const NEW_VISIT_SEGMENT = 'visitorType==new';
 
     /**
      * Return a single goal.
@@ -133,7 +135,7 @@ class API extends \Piwik\Plugin\API
      */
     public function addGoal($idSite, $name, $matchAttribute, $pattern, $patternType, $caseSensitive = false, $revenue = false, $allowMultipleConversionsPerVisit = false, $description = '')
     {
-        Piwik::checkUserHasAdminAccess($idSite);
+        Piwik::checkUserHasWriteAccess($idSite);
 
         $this->checkPatternIsValid($patternType, $pattern, $matchAttribute);
         $name        = $this->checkName($name);
@@ -187,7 +189,7 @@ class API extends \Piwik\Plugin\API
      */
     public function updateGoal($idSite, $idGoal, $name, $matchAttribute, $pattern, $patternType, $caseSensitive = false, $revenue = false, $allowMultipleConversionsPerVisit = false, $description = '')
     {
-        Piwik::checkUserHasAdminAccess($idSite);
+        Piwik::checkUserHasWriteAccess($idSite);
 
         $name        = $this->checkName($name);
         $description = $this->checkDescription($description);
@@ -268,7 +270,7 @@ class API extends \Piwik\Plugin\API
      */
     public function deleteGoal($idSite, $idGoal)
     {
-        Piwik::checkUserHasAdminAccess($idSite);
+        Piwik::checkUserHasWriteAccess($idSite);
 
         $this->getModel()->deleteGoal($idSite, $idGoal);
         $this->getModel()->deleteGoalConversions($idSite, $idGoal);
@@ -410,7 +412,7 @@ class API extends \Piwik\Plugin\API
 
         $segments = array(
             '' => false,
-            '_new_visit' => 'visitorType%3D%3Dnew', // visitorType==new
+            '_new_visit' => self::NEW_VISIT_SEGMENT,
             '_returning_visit' => VisitFrequencyAPI::RETURNING_VISITOR_SEGMENT
         );
 
@@ -518,18 +520,7 @@ class API extends \Piwik\Plugin\API
 
     protected function appendSegment($segment, $segmentToAppend)
     {
-        if (empty($segment)) {
-            return $segmentToAppend;
-        }
-
-        if (empty($segmentToAppend)) {
-            return $segment;
-        }
-
-        $segment .= urlencode(SegmentExpression::AND_DELIMITER);
-        $segment .= $segmentToAppend;
-
-        return $segment;
+        return Segment::combine($segment, SegmentExpression::AND_DELIMITER, $segmentToAppend);
     }
 
     protected function getNumeric($idSite, $period, $date, $segment, $toFetch)