+++ plugins/UsersManager/Model.php
@@ -291,6 +291,18 @@ class Model
         }
     }
 
+    public function removeUserAccess($userLogin, $access, $idSites)
+    {
+        $db = $this->getDb();
+
+        $table = Common::prefixTable("access");
+
+        foreach ($idSites as $idsite) {
+            $bind = array($userLogin, $idsite, $access);
+            $db->query("DELETE FROM " . $table . " WHERE login = ? and idsite = ? and access = ?", $bind);
+        }
+    }
+
     public function deleteUserOnly($userLogin)
     {
         $db = $this->getDb();