+++ plugins/UsersManager/tests/Integration/APITest.php
@@ -8,14 +8,114 @@
 
 namespace Piwik\Plugins\UsersManager\tests;
 
+use Piwik\Access\Role\View;
+use Piwik\Access\Role\Write;
 use Piwik\Auth\Password;
 use Piwik\Option;
 use Piwik\Piwik;
 use Piwik\Plugins\UsersManager\API;
+use Piwik\Plugins\UsersManager\Model;
 use Piwik\Plugins\UsersManager\UsersManager;
 use Piwik\Tests\Framework\Fixture;
 use Piwik\Tests\Framework\Mock\FakeAccess;
 use Piwik\Tests\Framework\TestCase\IntegrationTestCase;
+use Piwik\Access\Role\Admin;
+use Piwik\Access\Capability;
+
+class TestCap1 extends Capability
+{
+    const ID = 'test_cap1';
+
+    public function getId()
+    {
+        return self::ID;
+    }
+
+    public function getCategory()
+    {
+        return 'Test';
+    }
+
+    public function getName()
+    {
+        return 'Cap1';
+    }
+
+    public function getDescription()
+    {
+        return '';
+    }
+
+    public function getIncludedInRoles()
+    {
+        return array(
+            Admin::ID
+        );
+    }
+}
+
+class TestCap2 extends Capability
+{
+    const ID = 'test_cap2';
+
+    public function getId()
+    {
+        return self::ID;
+    }
+
+    public function getCategory()
+    {
+        return 'Test';
+    }
+
+    public function getName()
+    {
+        return 'Cap2';
+    }
+
+    public function getDescription()
+    {
+        return '';
+    }
+
+    public function getIncludedInRoles()
+    {
+        return array(
+            Write::ID, Admin::ID
+        );
+    }
+}
+
+class TestCap3 extends Capability
+{
+    const ID = 'test_cap3';
+
+    public function getId()
+    {
+        return self::ID;
+    }
+
+    public function getCategory()
+    {
+        return 'Test';
+    }
+
+    public function getName()
+    {
+        return 'Cap3';
+    }
+
+    public function getDescription()
+    {
+        return '';
+    }
+
+    public function getIncludedInRoles()
+    {
+        return array(Admin::ID);
+    }
+}
+
 
 /**
  * @group UsersManager
@@ -28,6 +128,11 @@ class APITest extends IntegrationTestCase
      * @var API
      */
     private $api;
+
+    /**
+     * @var Model
+     */
+    private $model;
     
     private $login = 'userLogin';
 
@@ -36,6 +141,7 @@ class APITest extends IntegrationTestCase
         parent::setUp();
 
         $this->api = API::getInstance();
+        $this->model = new Model();
 
         FakeAccess::$superUser = true;
 
@@ -224,6 +330,263 @@ class APITest extends IntegrationTestCase
         $this->assertEquals($expected, $access);
     }
 
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage UsersManager_ExceptionMultipleRoleSet
+     */
+    public function test_setUserAccess_MultipleRolesCannotBeSet()
+    {
+        $this->api->setUserAccess($this->login, array('view', 'admin'), array(1));
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage UsersManager_ExceptionNoRoleSet
+     */
+    public function test_setUserAccess_NeedsAtLeastOneRole()
+    {
+        $this->api->setUserAccess($this->login, array(TestCap2::ID), array(1));
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage UsersManager_ExceptionAccessValues
+     */
+    public function test_setUserAccess_NeedsAtLeastOneRoleAsString()
+    {
+        $this->api->setUserAccess($this->login, TestCap2::ID, array(1));
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage UsersManager_ExceptionAccessValues
+     */
+    public function test_setUserAccess_InvalidCapability()
+    {
+        $this->api->setUserAccess($this->login, array('admin', 'foobar'), array(1));
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage UsersManager_ExceptionNoRoleSet
+     */
+    public function test_setUserAccess_NeedsAtLeastOneRoleNoneGiven()
+    {
+        $this->api->setUserAccess($this->login, array(), array(1));
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage  UsersManager_ExceptionAnonymousAccessNotPossible
+     */
+    public function test_setUserAccess_CannotSetAdminToAnonymous()
+    {
+        $this->api->setUserAccess('anonymous', 'admin', array(1));
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage  UsersManager_ExceptionAnonymousAccessNotPossible
+     */
+    public function test_setUserAccess_CannotSetWriteToAnonymous()
+    {
+        $this->api->setUserAccess('anonymous', 'write', array(1));
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage UsersManager_ExceptionUserDoesNotExist
+     */
+    public function test_setUserAccess_UserDoesNotExist()
+    {
+        $this->api->setUserAccess('foobar', Admin::ID, array(1));
+    }
+
+    public function test_setUserAccess_SetRoleAndCapabilities()
+    {
+        $access = array(TestCap2::ID, View::ID, TestCap3::ID);
+        $this->api->setUserAccess($this->login, $access, array(1));
+
+        $access = $this->model->getSitesAccessFromUser($this->login);
+
+        $expected = array(
+            array('site' => '1', 'access' => 'view'),
+            array('site' => '1', 'access' => TestCap2::ID),
+            array('site' => '1', 'access' => TestCap3::ID),
+        );
+        $this->assertEquals($expected, $access);
+    }
+
+    public function test_setUserAccess_SetRoleAsString()
+    {
+        $this->api->setUserAccess($this->login, View::ID, array(1));
+
+        $access = $this->model->getSitesAccessFromUser($this->login);
+        $this->assertEquals(array(array('site' => '1', 'access' => 'view')), $access);
+    }
+
+    public function test_setUserAccess_SetRoleAsArray()
+    {
+        $this->api->setUserAccess($this->login, array(View::ID), array(1));
+
+        $access = $this->model->getSitesAccessFromUser($this->login);
+        $this->assertEquals(array(array('site' => '1', 'access' => 'view')), $access);
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage UsersManager_ExceptionAccessValues
+     */
+    public function test_addCapabilities_failsWhenNotCapabilityIsGivenAsString()
+    {
+        $this->api->addCapabilities($this->login, View::ID, array(1));
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage UsersManager_ExceptionAccessValues
+     */
+    public function test_addCapabilities_failsWhenNotCapabilityIsGivenAsArray()
+    {
+        $this->api->addCapabilities($this->login, array(TestCap2::ID, View::ID), array(1));
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage UsersManager_ExceptionUserDoesNotExist
+     */
+    public function test_addCapabilities_failsWhenUserDoesNotExist()
+    {
+        $this->api->addCapabilities('foobar', array(TestCap2::ID), array(1));
+    }
+
+    public function test_addCapabilities_DoesNotAddSameCapabilityTwice()
+    {
+        $addAccess = array(TestCap2::ID, View::ID, TestCap3::ID);
+        $this->api->setUserAccess($this->login, $addAccess, array(1));
+
+        $access = $this->model->getSitesAccessFromUser($this->login);
+
+        $expected = array(
+            array('site' => '1', 'access' => 'view'),
+            array('site' => '1', 'access' => TestCap2::ID),
+            array('site' => '1', 'access' => TestCap3::ID),
+        );
+        $this->assertEquals($expected, $access);
+
+        $this->api->addCapabilities($this->login, array(TestCap2::ID, TestCap3::ID), array(1));
+
+        $access = $this->model->getSitesAccessFromUser($this->login);
+        $this->assertEquals($expected, $access);
+
+        $this->api->addCapabilities($this->login, array(TestCap2::ID, TestCap1::ID, TestCap3::ID), array(1));
+
+        $expected[] = array('site' => '1', 'access' => TestCap1::ID);
+        $access = $this->model->getSitesAccessFromUser($this->login);
+        $this->assertEquals($expected, $access);
+    }
+
+    public function test_addCapabilities_DoesNotAddCapabilityToUserWithNoRole()
+    {
+        $access = $this->model->getSitesAccessFromUser($this->login);
+
+        $this->assertEquals(array(), $access);
+
+        $this->api->addCapabilities($this->login, array(TestCap2::ID, TestCap3::ID), array(1));
+
+        $this->assertEquals(array(), $access);
+    }
+
+    public function test_addCapabilities_DoesNotAddCapabilitiesWhichAreIncludedInRoleAlready()
+    {
+        $this->api->setUserAccess($this->login, Write::ID, array(1));
+
+        $access = $this->model->getSitesAccessFromUser($this->login);
+
+        $expected = array(
+            array('site' => '1', 'access' => 'write'),
+        );
+        $this->assertEquals($expected, $access);
+
+        $this->api->addCapabilities($this->login, array(TestCap2::ID, TestCap3::ID), array(1));
+
+        $expected[] = array('site' => '1', 'access' => TestCap3::ID);
+        $access = $this->model->getSitesAccessFromUser($this->login);
+
+        // did not add TestCap2
+        $this->assertEquals($expected, $access);
+    }
+
+    public function test_addCapabilities_DoesAddCapabilitiesWhichAreNotIncludedInRoleYetAlready()
+    {
+        $this->api->setUserAccess($this->login, Admin::ID, array(1));
+
+        $access = $this->model->getSitesAccessFromUser($this->login);
+
+        $expected = array(
+            array('site' => '1', 'access' => 'admin'),
+        );
+        $this->assertEquals($expected, $access);
+
+        $this->api->addCapabilities($this->login, array(TestCap2::ID, TestCap1::ID, TestCap3::ID), array(1));
+
+        $access = $this->model->getSitesAccessFromUser($this->login);
+        $this->assertEquals($expected, $access);
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage UsersManager_ExceptionAccessValues
+     */
+    public function test_removeCapabilities_failsWhenNotCapabilityIsGivenAsString()
+    {
+        $this->api->removeCapabilities($this->login, View::ID, array(1));
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage UsersManager_ExceptionAccessValues
+     */
+    public function test_removeCapabilities_failsWhenNotCapabilityIsGivenAsArray()
+    {
+        $this->api->removeCapabilities($this->login, array(TestCap2::ID, View::ID), array(1));
+    }
+
+    /**
+     * @expectedException \Exception
+     * @expectedExceptionMessage UsersManager_ExceptionUserDoesNotExist
+     */
+    public function test_removeCapabilities_failsWhenUserDoesNotExist()
+    {
+        $this->api->removeCapabilities('foobar', array(TestCap2::ID), array(1));
+    }
+
+    public function test_removeCapabilities()
+    {
+        $addAccess = array(View::ID, TestCap2::ID, TestCap3::ID, TestCap1::ID);
+        $this->api->setUserAccess($this->login, $addAccess, array(1));
+
+        $access = $this->getAccessInSite($this->login, 1);
+        $this->assertEquals($addAccess, $access);
+
+        $this->api->removeCapabilities($this->login, array(TestCap3::ID, TestCap2::ID), 1);
+
+        $access = $this->getAccessInSite($this->login, 1);
+        $this->assertEquals(array(View::ID, TestCap1::ID), $access);
+    }
+
+    private function getAccessInSite($login, $idSite)
+    {
+        $access = $this->model->getSitesAccessFromUser($login);
+        $ids = array();
+        foreach ($access as $entry) {
+            if ($entry['site'] == $idSite) {
+                $ids[] = $entry['access'];
+            }
+        }
+        return $ids;
+    }
+
     private function getPreferenceId($preferenceName)
     {
         return $this->login . '_' . $preferenceName;
@@ -232,7 +595,14 @@ class APITest extends IntegrationTestCase
     public function provideContainerConfig()
     {
         return array(
-            'Piwik\Access' => new FakeAccess()
+            'Piwik\Access' => new FakeAccess(),
+            'observers.global' => \DI\add([
+                ['Access.Capability.addCapabilities', function (&$capabilities) {
+                    $capabilities[] = new TestCap1();
+                    $capabilities[] = new TestCap2();
+                    $capabilities[] = new TestCap3();
+                }],
+            ]),
         );
     }
 }