+++ CHANGELOG.md
@@ -10,6 +10,13 @@ The Product Changelog at **[matomo.org/changelog](https://matomo.org/changelog)*
 
 * Added new event `API.addGlossaryItems` which lets you add items to the glossary.
 * There is a new widget metadata property, `isReusable`, that marks whether a widget can be added multiple times to a dashboard. It is up to the consumer of widget metadata to respect this property.
+* A new role has introduced called "write" which has less permissions than an admin but more than a view only user.
+* Added new API method `UsersManager.getAvailableRoles` to fetch a list of all available roles that can be granted to a user.
+* Added new API method `UsersManager.getAvailableCapabilities` to fetch a list of all available capabilities that can be granted to a user.
+* Added new API method `UsersManager.addCapabilities` to grant one or multiple capabilities to a user.
+* Added new API method `UsersManager.removeCapabilities` to remove one or multiple capabilities from a user.
+* The API method `UsersManager.setUserAccess` now accepts an array to pass a role and multiple capabilities at once.
+* Plugin classes can overwrite the method `requiresInternetConnection` to define if they should be automatically unloaded if no internet connection is available (enable_internet_features = 0)
 
 ### Breaking Changes
 * Changed some menu items to use translation keys instead (see [PR #12885](https://github.com/matomo-org/matomo/pull/12885)).